+++ src/lib/input-validation/constant/errorType.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019. AccelByte Inc. All Rights Reserved
+ * Copyright (c) 2019-2022. AccelByte Inc. All Rights Reserved
  * This is licensed software from AccelByte Inc, for limitations
  * and restrictions contact your company contract manager.
  */
@@ -19,6 +19,7 @@ export const CommonValidationErrorType = Enum(
   "notContainsUpperCaseAndLowerCase",
   "containsTwoCharactersInArrow",
   "containsBadWords",
+  "containsForbiddenWords",
   "containsDecimal",
   "invalidFileExtensions",
   "exceedMaximumFileSize",