+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.34.0]
+### Added
+- Default rule for custom email validation
+-  Add new validation `validateForbiddenWords` to validate whether there is a forbidden word in a string
+
 ## [1.33.5]
 ### Added
 - Error translation for `10142` and `10143` in `adminServiceErrorTranslationMap`